+++ repository/config.go
@@ -4,10 +4,12 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"github.com/cloudogu/ces-commons-lib/dogu"
-	"github.com/cloudogu/k8s-registry-lib/config"
 	"reflect"
 	"strings"
+
+	"github.com/cloudogu/ces-commons-lib/dogu"
+	"github.com/cloudogu/k8s-registry-lib/config"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type configName string
@@ -53,6 +55,7 @@ func (cr configRepository) get(ctx context.Context, name configName) (config.Con
 		cfgData,
 		config.WithPersistenceContext(getPersistentContext(cd.rawData)),
 		config.WithInitialListResourceVersion(listResourceVersion),
+		config.WithLastUpdated(cd.lastUpdated),
 	)
 
 	return cfg, nil
@@ -79,6 +82,7 @@ func (cr configRepository) create(ctx context.Context, name configName, doguName
 	}
 
 	cfg.PersistenceContext = resource.GetResourceVersion()
+	cfg.LastUpdated = getLastUpdated(resource)
 
 	return cfg, nil
 }
@@ -96,8 +100,17 @@ func (cr configRepository) update(ctx context.Context, name configName, doguName
 	}
 
 	cfg.PersistenceContext = resource.GetResourceVersion()
+	cfg.LastUpdated = getLastUpdated(resource)
 
 	return cfg, nil
+}
+
+func (cr configRepository) setOwnerReference(ctx context.Context, name configName, owners []v1.OwnerReference) error {
+	_, err := cr.client.SetOwnerReference(ctx, name.String(), owners)
+	if err != nil {
+		return fmt.Errorf("could not set owner Reference: %w", err)
+	}
+	return nil
 }
 
 func (cr configRepository) saveOrMerge(ctx context.Context, name configName, cfg config.Config) (config.Config, error) {
@@ -142,6 +155,7 @@ func (cr configRepository) saveOrMerge(ctx context.Context, name configName, cfg
 	updatedConfig := config.CreateConfig(
 		updatedRemoteConfigData,
 		config.WithPersistenceContext(getPersistentContext(updatedResource)),
+		config.WithLastUpdated(getLastUpdated(updatedResource)),
 	)
 
 	return updatedConfig, nil
@@ -238,8 +252,12 @@ func createConfigWatchResult(lastCfg config.Config, result clientWatchResult, co
 
 	return configWatchResult{
 		prevState: lastCfg,
-		newState:  config.CreateConfig(cfgData, config.WithPersistenceContext(result.persistentContext)),
-		err:       nil,
+		newState: config.CreateConfig(
+			cfgData,
+			config.WithPersistenceContext(result.persistentContext),
+			config.WithLastUpdated(result.lastUpdated),
+		),
+		err: nil,
 	}
 }
 
@@ -247,7 +265,7 @@ func getPersistentContext(rawData any) string {
 	switch r := rawData.(type) {
 	case string:
 		return r
-	case resourceVersionGetter:
+	case resourceMetaAccessor:
 		return r.GetResourceVersion()
 	default:
 		return ""