+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#28] set owner reference of config map and secrets of dogu config
+- [#28] set owner reference of dogu descriptor config map
 
 ## [v0.5.1] - 2024-11-19
 ### Changed